Externalize KeyStore usage
